third_party/WebKit/LayoutTests/css3/device-adapt/viewport-initial-width-zero-expected.txt

ALERT: viewport size 100x200 scale 0.25 with limits [0.25, 5] and userScalable true
